User Progress backend controller service response and fixes

+++ spring-boot-backend/src/main/java/com/example/springbootbackend/repository/ProgressRepository.java
@@ -4,7 +4,7 @@ import com.example.springbootbackend.model.Progress;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface ProgressRepository extends JpaRepository<Progress, Long> {
-    List<Progress> findByUserId(Long userId);
+public interface ProgressRepository extends JpaRepository<Progress, Integer> {
+    List<Progress> findByUserId(Integer userId);
 }
